#0
Updated gridftp-simple-probe to use UberFTP instead of globus-job-run
to cleanup.  This should fix the problem at UNL where they do not 
have a CE on their head node.
Scot, if you are reading this, I now need to add UberFTP to the RSV 
pacman dependencies.  Also, bin/metrics should be updated with this new copy.



git-svn-id: https://vdt.cs.wisc.edu/svn/rsv/branches/rsv-storage@11236 4e558342-562e-0410-864c-e07659590f8c